tdf#114068 allow to translate media control tooltips in navigator pane
tooltips for play/pause/stop/repeat/mute were explicitly flagged as
non-translatable for some reason. There is a media toolbar that has the
same set of controls and there could be a way to share the strings, but
they only are a handful, so not really worth a big refactoring of the UI
files for that.

Change-Id: Ia0e8e2bd664916392049df5cd537d51eb04005d1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192873
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Tested-by: Jenkins
(cherry picked from commit 74e55a3c756c1f3ce540bf7ab575ca40a334172d)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192911
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>