sc: sheet geometry and header invalid. needs the actual tab
This fixes row/column header invalidation when using the sheet
view. Previously, because the wrong tab number was passed, the
invalidation was not performed and the header didn't reflect the
cells. This was especially apparent when changing the auto-filter.

In the change all the sheet geometry and header invalidations are
changed so it should also fix some other - not yet noticed bugs.

Change-Id: Ia382882fedce18750dfc812793501573fff19093
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192887
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192996
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>