tdf#166803 – wsl-as-helper: fix cross-compilation to aarch64
649314399483331244538128149ebd0556166540 did fix building on aarch64
natively, but in turn broke cross-compilation

the problem in tdf#166803 is reproducible with VS2019, but not with
VS2022, but that cannot be installed on our old baseline and in turn
new baseline is using the wsl-as-helper method.
So the patch is needed on the branch as well.

Change-Id: I1c9a4eabe1489b3c4bab5c07cd1e62fe29992530
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192910
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
(cherry picked from commit 7855c25da38686ddcccb7feb8410f04ed620bc34)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/197803