tdf#168970 Remove {OFieldDescControl,OTableFieldControl}::dispose
The dispose methods were only called in the destructors of the these two
classes so they didn’t work like a normal dispose method to break
reference cycles. Other than that the only useful work they did was to
hide all of the controls in OFieldDescControl. This turns out to be kind
of dangerous because if one of the controls had the keyboard focus then
it would end up sending the focus lost event which would in turn call an
undefined pure virtual function and crash.

This patch is needed to fix the crash when OFieldDescControl is used in
the new table wizard and the copy table wizard. In that case we don’t
need to try to flush the modifications because the new table is just
discarded as if the dialog was cancelled when the window is closed.

Change-Id: I56b585d2dcd5bbf2ee2f0d770a33a02685e60d0c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192977
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>