Fix renaming field type in InsDeletedFieldType
The function has this comment:

  // - If the same type is found, the deleted one has to be renamed.

And obviously, the loop in it intends to do that. It searches the name;
when it finds one, it uses a counter to add a suffix to find an unused
name; and finally, it assigns the found unused new name.

The problem was, that it assigned it to a local variable, not to the
field type.

This happened in commit 263153842741d7ce21cc0bf1c5296a55a1138024
(String to OUString, 2013-07-16). Before that, SwFieldType::GetName
returned a const reference to a string; and that string was the field
actually representing name in the relevant field types. The function
InsDeletedFieldType took the name by const reference, and const_casted
it when renaming. The mentioned commit changed SwFieldType::GetName to
return by value, meaning that InsDeletedFieldType operated on a local
variable since then.

Later commits changed the local variable in InsDeletedFieldType to be
non-reference, and later the name in SwFieldType became a UIName.

This change introduces a normal setter for the field type name, to
avoid the hackish rename method, that led to this problem in the first
place.

This is a problem found by code reading. I don't know a user-visible
problem caused by this, so no unit test.

Change-Id: Ibe6fb1e0a11c5193cc253efe7241592d7ce9ca77
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192931
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>