tdf#164718: allow splitting multiline fields in SwTextFrame::FormatAdjust
Commit 167be9363ea505b334aa595b273707d7d9217863 (tdf#160526, tdf#160549:
fix split conditions at page start, 2024-04-11) had changed a condition
to split at a line, disallowing it when the position stored in nEnd is
equal to the start of the line. It turned out, that the multiline field
portions are special; for all following portions, the line start is the
position after the field placeholder character, and the length is zero.
This change handles such fields into account.

To check if this is a multiline field, it calls SwLineLayout::IsRest.
The respective SwLineLayout::SetRest is only called in two places in
SwTextFormatter::WhichFirstPortion. SwTextFormatInfo::SetRest is also
called there, with nullptr as the argument, which explains why
SwTextFormatInfo::GetRest didn't help in the pre-existing code in
SwTextFrame::FormatAdjust.

The split of the field is correct here. There is a different problem,
that this split is undone later at some place, if widow/orphan control
is not 0; that should be fixed separately.

Change-Id: I3141a4ad4c1e8b3f5437d1fa63f0fd04cffcb297
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/192978
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>