tdf#101630: Do not prompt twice for OAuth2 authentication
The problem was:
When adding a service for Google Drive, clicking OK in the File
Services dialog would initiate the OAuth2 authentication procedure;
after successful confirmation, immediately another authentication
procedure started.

The reason was, that the first procedure was initiated by the code
in RepoContent::getRepositories, which created a session, but did
not register it, nor saved its token. Then, in Content::getSession,
another session had to be created, which had to authenticate again.

Before this change, the registration for an OAuth2-authenticated
session used a key including the repo. Therefore, RepoContent did
not know which repo it had to use for registration.

This change unifies the previously duplicated code used to create
a session. For OAuth2-based sessions, the key for registration is
now only the service's binding URL, which allows to use the same
key both in RepoContent::getRepositories and Content::getSession.
Now the successfully created session in getRepositories is not
deleted, allowing it to be registered and reused.

Also, after the OAuth2 authentication failed (which means that it
was cancelled by the user), the password dialog is not shown.

Change-Id: Ic84a8e034eb0900aca3b4cac0d73dd7443d6380d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193002
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>