tdf#166488 Fix local variable used to store char from SvxCharacterMap
The SvxCharacterMap dialog is used to select a character and returns a
32-bit Unicode codepoint via GetChar. When invoked from
SvxBulletPickTabPage this was then being stored in a sal_Unicode
variable so it would get truncated to 16-bit. This patch just fixes it
to store the result in a sal_UCS4.

Change-Id: I49984817388dc8d77284f18c085804bed049d740
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193353
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
(cherry picked from commit 4888f6f3f59785d0e661b58060bd8458f866bf10)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193365
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>