tdf#169090: Only change bookmark type to NUMITEM for paragraph PaMs
In a real-world DOCX, using redlines, there was a bookmark with a name
starting with "__RefNumPara__", positioned not in the para properties,
but in the middle of a para text. Opening that DOCX in my debug build
failed an assertion in ctor of CrossRefBookmark, which requires that
IDocumentMarkAccess::IsLegalPaMForCrossRefHeadingBookmark returns true.

It makes sense that a NUMITEM bookmark expects a paragraph-wide PaM, so
just reorder the checks in SwXBookmark::attachToRangeEx. Now it works
for NUMITEM the same way as for HEADING.

Change-Id: Ia33b54f138d7f61ee72dc3b5c270efdd0fec8068
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193024
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 0b3a4ee85ae264d96bb6dc221dbcf4c394108bff)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193039
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>