add comment: 'vcl PrinterController: oversized? align top, not center'
This was not a hard-and-fast calculation
about where the top should start,
so document this using arbitrary terms.

Elsewhere in the code we can see
that bFullPaper can shift by PageOffset,
and since all the content is not going to fit anyway,
just avoid any potential shift
(since documents typically include a significant top margin anyway).

bFullPaper is normally set to true (even on pages with empty content).

Change-Id: I43a83e25a251cc6b66dfa46d7585332041ec0bf8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193048
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins