mso-test: emit w:settings elements in correct places
When loading and then saving the document from ooo#120405 ,
we end up with incorrect ordering of some of the settings sub-elements.

Re-structure the code here to match the schema ordering.
Pull code related to each sub-element closer together,
to make it easier to see when we need to emit an element.

Change-Id: I4ef7c2f275911f1226b3c64fe12605de84cb418d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193083
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193223
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193242