tdf#127201 extension mgr: Extract helper for calculating entry rect
Implement a helper method ExtensionBox::GetEntryRect
that can be used to calculate the rectangle of any index
and use it in two places that were previously duplicating
similar logic.

The new method could also be used to implement the
comphelper::OAccessible::implGetBounds override
in case of a future a11y implementation for the entries
in the extension manager.

Change-Id: I6c6d16564db6f20f3d77b65308d5c07b98a41006
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193107
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins