qt a11y: Report orientation
If either the HORIZONTAL or VERTICAL a11y state
are set, report the corresponding
QAccessible::Attribute::Orientation attribute newly
introduce in Qt's qtbase commit [1]

    commit b53656bfa70461a643ff564ff33eb2c049316e15
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Oct 1 14:59:07 2025 +0200

        a11y: Introduce accessible attribute for orientation

This e.g. shows the corresponding horizontal or
vertical AT-SPI2 states in Accerciser's "Interface Viewer",
section "States" for the horizontal and vertical toolbars
in Writer when running with the qt6 VCL plugin and selecting
the corresponding objects in Accerciser's treeview of the LO
a11y hierarchy.

[1] https://code.qt.io/cgit/qt/qtbase.git/commit/?id=b53656bfa70461a643ff564ff33eb2c049316e15

Change-Id: I03f062250949e0f6f8bd1717e2eeeb85790975bc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191734
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>