Add an ID for OTableEditorCtrl and SbaGridControl for UITests
This makes it easier to find the two controls in UITests. Otherwise they
are buried in a hierarchy of windows that also don’t have a name so they
are difficult to get access to. I mainly wanted to do this to switch
between the selected rows via key presses. There doesn’t seem to be any
commands or any other easy way to select a row in a test.

Change-Id: Ic20ee4f2b1e416e4274a2f159a27448f4b3580ef
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193173
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
Tested-by: Jenkins