sw: reduce duplication of w:num items
When saving to .docx format we create new w:num entries for lists even
if equivalent entries already exist. This creates a lot of bloat in big
documents that have heavy use of numbered lists. A document that can be
saved with ~150 num entries can end up with ~2000. This causes
compatibility issues with other software that doesn't expect such high
number of numbering styles.

Change-Id: Ie8667c0a6d98a77c7c429a9a2925d31ee19fd31c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193164
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>