tdf#130857 Define widgets in Conditional Formatting dlg more in order
Switch the order in which the widgets in the .ui file
are defined so that the order better matches the visual appearance
and logical order, which makes sure that tab focus order with the
Qt-based VCL plugins is correct as well when using native Qt
widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where
focus order is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

More tweaking may be needed, but this at least
already improves the keyboard focus order with
the qt6 VCL plugin for the dialog newly supported
since

    Change-Id: I604295dee25066373548c466c8c4f277feebeb54
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 30 15:27:40 2025 +0100

        tdf#130857 tdf#169006 qt weld: Support Calc "Conditional Formatting" dialog

Change-Id: Ic0a4e6e83feccfeed934744d5cabdb6d3de1c439
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193216
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins