we dont know the 'family type'
its certainly not whatever arbitrary FamilyType is set on
the default font of the OutputDevice.

DONTKNOW is better, that way at least the other properties that we do
know will match during selection of a matching font when creating
embedded font output.

Especially on windows where the font in the pdf embedding test is
detected as "FAMILY_MODERN" by the windows backend, while the default
font is typically FAMILY_ROMAN and so doesn't match exactly, so the
next best match was fairly arbitrary and equal weight given to
family/weight/pitch so the non-bold font got used for the bold font.

Change-Id: If16b2851cb9461969409fa58eb43430516ed0a8b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193231
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>