BrowseBox: Don’t set last column width to LONG_MAX when win is invisible
BrowseBox::AutoSizeLastColumn calls SetColumnWidth with the width set to
LONG_MAX. SetColumnWidth then tries to limit the actual width to the
remaining space for the last column. However, this was only done when
the window is visible, presumably because the remaining space
calculations might not work otherwise. This width is later used to
calculate a rectangle starting from the column’s X position. This of
course ends up overflowing the Long variable so it could be counted as
undefined behaviour. Instead, let’s just bail out early from the
function when the window isn’t visible and leave the width as it was.
The actual width will be recalculated when the window becomes visible
anyway because Window::Show calls Resize.

This was picked up by Stephan Bergmann with UBSAN and it was discussed
in the comments for this patch:

https://gerrit.libreoffice.org/c/core/+/193174

Change-Id: Ic2922a3d98c8ebc2e7f2d04c8c59105fdbd4e666
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193268
Reviewed-by: Stephan Bergmann <stephan.bergmann@collabora.com>
Tested-by: Jenkins