tdf#169151 Allow modifying accessible text alternative in sidebar
This patch adds a 'Text Alternative...' menu item to the Writer
Navigator context menu for content type entries that can have this
(Frames, Images, OLE objects, and Drawing objects). It also
adds an action button to the functions toolbar in the Writer Navigator
tool box for these content types.

Seems gtk isn't able to use the 'action-name' property in the .ui
file. <property name="action-name">.uno:ObjectTitleDescription</
property> works for qt and gen, which allows to not need the alias
"TargetURL" UNO commands added to GenericCommands.xcu and the icon
declarations added to the icon-theme link text files. I think it would
be nice if gtk supported this .ui property as well.

Change-Id: I519590fc56f39e9c6c2b2fccea568271e5fd3e77
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193376
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>