lok: FormattedField: ensure formatter has min and max
If not use the limits of a 32 bit integer. Without this check if min/max
are not set already, the json min/max will be set to 0/0 which is
problematic in web clients where it rejects all inputs.

Signed-off-by: Dennis Francis <dennis.francis@collabora.com>
Change-Id: I9fa2bb68b47626ad42bae51e52e4b14baecb82fb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193259
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
(cherry picked from commit f62a357d7b32084de1e32e88d599938a59b94f11)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193757
Tested-by: Andras Timar <andras.timar@collabora.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>