text in an fodg converted from pdf reimporting split over two lines
when it began as one line.

The original pdf breakup sets a font size which when exported and
reimported through the various intermediate conversions results in a
larger size used on reimport that doesn't fit in the original textbox.

Best thing is to reduce to what can be represented as a twip and convert
from that starting point.

Change-Id: I24a6ccb4977112f2dc708ba737be07aeb86437dc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193273
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>