tdf#169158: Call FormatObjsAtFrame again, when at-page fly needs that
Formatting a cell with non-default vertical alignment is done in two
steps. In the end of the first step, SwContentNotify::ImplDestroy
detects the difference of size/position of the table content, and the
non-default vertical alignment, and invalidates the cell and its page.
Then the next formatting would correct the size/position mismatch; it
obviously relies on the second pass of formatting.

In SwLayAction::InternalAction, the formatting loop did a single call
to SwObjectFormatter::FormatObjsAtFrame, which formatted at-page fly
objects, and then an inner loop which formatted the rest of the page
content. In that inner loop, it performed validation of all flags,
then called FormatContent, and checked the validity. That meant, that
any invalidation that could happen in at-page objects formatting time
would be lost. Then, there was nothing to indicate that the at-page
objects could need a second formatting attempt; the outer loop could
only run once, even when the at-page formatting needed another run.

This change introduces a dedicated flag for at-page fly invalid state.
It is checked in SwLayAction::InternalAction in a new loop, to make
sure that SwObjectFormatter::FormatObjsAtFrame is repeated as needed.

Change-Id: Id6f4fef10cc031af95c38f0074178405027c513d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193298
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>