related tdf#163347: Sync colour table to chart colors list selection
This patch makes the selected data series entry color in the chart
colors list be selected in the color table for SvxDefaultColorOptPage.
The color selection made in the color table could be from a different
palette than the original. The first palette in the list of palettes
that has a matching color is choosen.

I noticed when a color selected from the Material palette was set for a
data series color, the Tonal palette was shown when reselecting it. An
off by 1 index bug that causes the last palette in the list of
palettes to not be searched for the color was found.

Another thing I noticed is when a data series chart color was set to a
color in the Freecolour HCL palette that needed to be scrolled to see
it, it didn't do it. Commenting out mbScroll in the condition in the
SelectItem function that checks its value solved the problem. I don't
get the need for the ValueSet mbScroll variable. The only place its
value can change is in the Format function. That doesn't get called in
the SelectItem function unless mnCols is 0. The SelectItem function is
the only place its value is checked. I don't see a problem with always
checking if the visible area needs to be scrolled. This patch removes
the ValueSet mbScroll variable.

Change-Id: I1f05a3ad8a0b64b2a28d3001cace391379cc00c6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193437
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>