mso-test: duplicate sz element in styles.xml
This is using the test document from tdf94682-1.odt.
We want to write w:sz in multiple places, which creates duplicate elements,
which officeotron rules invalid.
Use our existing strategy of collecting data and then writing it at the
end.
This is not ideal, because who is say which one of the font-sizes is
the most correct? But our existing unit tests seem to believe that the
"last" writer of the value is correct, so this approach works.

Change-Id: I5f720f0d2c5d3efb7c2aec5ec955cc761fbb980d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193341
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>