Fix indentation in the UI tests
A bunch of the UI tests were using only three spaces to indent the
“def” line for a method. In most of the tests, the subsequent lines were
indented by five spaces so that most of the code ends up on a multiple
of 4 spaces and it works out. A few of the tests only indented the
subsequent lines by one additional space so if we fix the “def” line we
need to also fix the rest of the function because otherwise it would end
up having the same indentation as the declaration.

This patch normalises the indentation to use multiples of 4 spaces. I
think the change is worthwhile because the tests are often copy-pasted
and the misindentation is probably spreading that way.

The change was made with this snippet of Emacs lisp:

(with-temp-buffer
  (call-process "bash" nil (current-buffer) nil
                "-c" "git ls-files | grep '/uitest/.*\\.py$'")
  (goto-char (point-min))
  (while (re-search-forward "^.+$" nil t)
    (save-excursion
      (find-file (match-string 0))
      (when (local-variable-p 'require-final-newline)
        (setq require-final-newline nil))
      (goto-char (point-min))
      (while (re-search-forward "^   [^ ]" nil t)
        (goto-char (match-beginning 0))
        (let ((end (if (looking-at " *def\\_>")
                       (copy-marker (save-excursion (end-of-defun) (point)))
                     nil)))
          (insert " ")
          (when end
            (forward-line)
            ;; Find the next indented line
            (when (re-search-forward "^\\( +\\)[^ ]" end t)
              (let ((indent (length (match-string 1))))
                ;; If it’s not indented by 8 then adjust the whole
                ;; function by the right amount
                (when (/= indent 8)
                  (beginning-of-line)
                  (increase-left-margin (point) end (- 8 indent)))))
            (set-marker end nil)))))))

Change-Id: Iaec817dd5b982010f460be61950003b6ec8e42b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193371
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>