sw: fix SwDoc::IsInHeaderFooter() with flys
If the node is in a fly, this can only print:

  sw/source/core/doc/doclay.cxx:1619: Found a FlySection but not a Format!

This is a regression from commit
e07feb9457f2ffb373ae69b73dda290140e4005f which used the wrong node's
anchored objects: of course the fly is never anchored to itself, so this
will never find anything.

It turns out that there is already a function SwNode::GetFlyFormat()
that can do most of the work, even using the layout to speed it up, so
just use that to get the anchor.

Change-Id: I7a22231f929175ed7c5c11724882b03890781cfc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193476
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193531
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>