tdf#164070 always install en-US UI localization on windows
if en-US is not available, and the user didn't manually configure the UI
language, adding another language (via means of mar updater for example)
can cause that new language being used instead of the language the user
had installed initially since there's no explicit default/fallback.

Since any language that will be added in future is likely to not use a
latin script/will be quite unfamliliar to the user, having en-US to rely
on as a fallback is a much better experience.

The problem occurs on systems where the selected UI language is not
matching the windows language, so for new installations the user has to
go to options and set the language manually. That in turn then will be
an additional safeguard for further updates, since now there's a
language explicitly configured and thus no reason for LO to "guess" what
to use as default.

Change-Id: I67dd9a9ddd9de3ce79eb9144eab06b6f10134c57
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193422
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>