tdf#169006 sc a11y: Keep keyboard focus in conditional formatting dlg
In ScCondFormatList::AfterTypeListHdl, the existing
ScCondFrmtEntry can be deleted, and with it, its combobox
indicating the type, and a new ScCondFrmtEntry created
to replace it.

In the case that combobox was focused (which is usually
the case, because ScCondFormatList::AfterTypeListHdl
is the handler that gets called when the active entry
changes), then make sure that its combobox for the
type takes over/receives keyboard focus, so the user
can continue in the "same" (from their perspective)
place.

While the case of gtk3 with native GTK widgets and
qt6 with SAL_VCL_QT_USE_WELDED_WIDGETS=1 to use native
Qt widgets would at least allow to move the keyboard
focus to somewhere in the dialog using the Tab key
previously, the vcl implementation would not allow
this, breaking keyboard accessibility of the dialog
altogether.

Change-Id: Iaa3e087cda078678d1f160ce8a49938b27e6528c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193431
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
(cherry picked from commit 752adf6cc9d869bbcf34fcf09c80da502398cb81)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193432
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>