tdf#136220 sw: text formatting: fix breaking line with multiple fields
... when there is an anchored object in the footer, and the last field
fits only partially on the line.

This stopped working with commit 8e3afdb5989d571410350f1d43fcf26492a4eaff
removing the IsFooterFrame() check in SwTextFly::GetTop().

Previously a new line below the last line that fits on the page started
with the field rest portion, when the not-fitting line is deleted in
SwTextIter::TruncLines() this caused pDel->IsRest() to be true and
setting SetFollowField(true) which later caused the mbFieldFollow to be
set in the follow text frame.

Now the pDel->IsRest() is false because the line contains only a big fly
portion due to the fly in the footer, and then SetFieldFollow(false) and
the follow doesn't start with the rest of the field.

There is a long standing bug in lcl_NeedsFieldRest() causing this, it
simply doesn't work if there are multiple fields in a line, it
erroneously considers the first field to be the last one in the line.

Change-Id: I99e9f68e8e9f8019e4d22ce5d7874d4cdac285e5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193474
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Tested-by: Jenkins
(cherry picked from commit d2c387bc609a235406b5c6c434ac4cc06a268513)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193528
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>