tdf#169253: show infobar when changes are tracked in Calc
There are several differences in Calc compared to Writer with regards to
change tracking. First, there is no Track Changes toolbar there. Second,
it's impossible to have changes without tracking. And third, it has more
detailed control over which changes are shown, e.g. by date, author, etc.,
which makes it possible to have the changes "shown", and still have all
existing changes hidden.

Because of the last detail, I decided to show the infobar regardless of
the "show" flag. The HiddenTrackChanges registry configuration allows to
disable it, if needed.

Change-Id: I3dcfeb68d13791e383665b5b769a1375ba3386f0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193624
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>