SfxObjectShell::SaveTo_Impl may overwrite its original output file
at which point the cached value of the 'Modified' timestamp is that
of the pre-overwrite time. Drop the cached value and fetch it again
on next demand.

Otherwise another overwrite#2 attempt will compare against the
timestamp of the pre-overwrite-#1 and not of the post-overwrite#1
and a 2nd overwrite behaves differently that a 1st overwrite.

It assumes that the timestamp mismatch means the file has been modified
by an external event and requests an XInteractionHandler intervention to
approve the 2nd overwrite (but not the 1st overwrite).

This is indistinguisable from the 1st overwrite if there is no XInteractionHandler
during the 2nd overwrite, but not if there is an XInteractionHandler
set.

The QuietInteractionHandler, used for headless and silent mode, defaults to
aborting the save for queries it doesn't recognize, so will throws an exception in
this case.

The result is that CppunitTest_sfx2_misc 'testOverwrite' fails if the
same interaction handler present during the 1st overwrite is still
present during the 2nd overwrite. Which the next commit will introduce.

Change-Id: I80b997ce6adba38049c0e92bd8b18d160a0a2292
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193520
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>