tdf#169275 sd UI: clear direct format when turning off bullet/num
Open the bugdoc, go to the last paragraph of the only shape in the
document, press 'Toggle Unordered List', bullet disappears, press it
again, bullet appears and the indent is 0cm, even if the master page
defines Level 1 to have 0.30cm indent and Level 2 to have 1.5cm indent,
so we end up with a third, unexpected indent.

What happens is that the last paragraph is technically level 3, but it
has direct formatting to look like level 2. Turning off bullets resets
the level to 0 (no bullets), but the numbering rules as direct format is
still there. Once the bullet is toggled on, the level 1 indent of this
numbering rules gets applied, which is confusing and is inconsistent
with the numbering rules from the master page.

Fix the problem by removing the numbering rules from the selected
paragraph(s) when you toggle off a bullet or numbering. The numbering
level will be lost anyway, and this gives "toggle on" a chance to again
connect to the outline style of the master page, resulting in a
consistent indent of Level 1 (0.30cm in the bugdoc case).

This keeps the "to numbering" case unchanged, which has to be direct
formatting (for the bugdoc case), since the master page declares
bullets. This only changes editing for the non-masterpage case, since
TextObjectBar::ExecuteImpl() won't dispatch FN_SVX_SET_BULLET for master
pages. Finally, only do this for outline shapes, leave the behavior
unchanged for other shape types, which are not connected to master page
styles.

Change-Id: I4df409e450862fc472d115d196c47fa2d7f17c7b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193521
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>