readlicense_oo/docs/readme.xrm: Update GTK/ATK min version
... after

    commit c392f7b8a614c7ef32cfa9888aeb06b27b535390
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Nov 5 17:44:57 2025 +0100

        gtk3: Require GTK >= 3.24

Thanks to Ilmari for pointing it out in [1].

Also, mention ATK as a dependency along with the GTK 3
version, because ATK is used by the GTK library and
the gtk3 VCL plugin.

Drop the explicit mention of at-spi2. Except for the
build-time gtk3 a11y tests, libatspi isn't used directly,
but only indirectly via ATK or on the "receiving end" by
assistive technology. But then, it's the responsibility
of the assistive technology or desktop environment
packaging to ensure that libatspi is installed.

While at it, also change ""KDE" to "KDE Plasma", as
that's the name of the desktop environment, and
use the term "desktop environment" instead of "GUI".

[1] https://gerrit.libreoffice.org/c/core/+/193477/comments/86c382ee_b979934f

Change-Id: Ic93234b8b9bee0bd770ae99429fd2ed9fe6ae767
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193526
Tested-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>