tdf#169291: Fix rendering of solid grid in Writer.
When solid lines are used for rendering the grid,
different input agruments are expected by DrawGrid() method.
The start position of the lines is needed, which is determined
by the edge of the drawing area.

Also move the DrawGrid() call out of the loop. It's enough to call
the rendering code once for the vertical and once for the horizontal
lines.

Change-Id: I1f1eddd2c3c9abd91fbd3e48eaee9bd10385ef6e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193541
Reviewed-by: Tamás Zolnai <zolnaitamas2000@gmail.com>
Tested-by: Jenkins