mso-test: invalid xdr:twoCellAnchor
This is using the test document from fdo73254-1.xls.

Importing and exporting to XLSX and then running officeotron generates an error.

The xdr:twoCellAnchor has mandatory subelements, so we cannot skip writing them,
even if we have no graphic or media URL.

The code I am removing seems to be from
    commit cbcffbaee204e9dfbdc2125dda73e36d77e9b844
    Author: Jan Holesovsky <kendy@collabora.com>
    Date:   Tue Sep 5 16:53:14 2017 +0200
    tdf#106867: Export videos in PPTX.

Change-Id: Ie095bad8ea9bf92b553448205f00c9cacf42d513
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193538
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>