tdf#169311: Improve accuracy of rendering of grid points in Writer.
Before: The first visible grid point was used as anchor point which could
be a main grid point or a secondary grid point. However because of
the inevitable numeric imprecisions, different anchor points lead
to different final position values (one pixel differences).

After: A main grid point is used as anchor point, whose position can be
determined precisely. Whichever grid points are actually visibile in the
current painting area the selection of anchor point is consistent and so
the final position value will be too.

Change-Id: I138fcb4e018b94d4d8c3832179fda72091268e1a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193578
Tested-by: Jenkins
Reviewed-by: Tamás Zolnai <zolnaitamas2000@gmail.com>